% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_tag}
\alias{case_tag}
\title{case_tag}
\format{
Dataframe with covariate data.
\describe{
\item{code_tag}{Covariate data for the 104 breast cancer samples and 17 healthy tissue samples}
...
}
}
\usage{
case_tag
}
\description{
A data.frame containing covariate data for the 104 breast cancer samples and 17 healthy breast tissues.
}
\keyword{datasets}
