% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_full_data}
\alias{check_full_data}
\title{check_full_data}
\usage{
check_full_data(full_data, na.rm = TRUE)
}
\arguments{
\item{full_data}{Matrix with the columns of the input matrix
corresponding to the individuals belonging to the level.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.
If \code{FALSE}, an error occurs in case of \code{NA} rows.}
}
\value{
Return \code{full_data} without NAN's and as a matrix
}
\description{
Checking the full_data introduces in the package
}
\examples{
\dontrun{
check_full_data(full_data, na.rm = TRUE)}
}
