% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_gene_selection}
\alias{check_gene_selection}
\title{check_gene_selection}
\usage{
check_gene_selection(num_genes, gen_select_type, percent_gen_select)
}
\arguments{
\item{num_genes}{Number of genes in the full_data}

\item{gen_select_type}{Type of gene selection to be used. Choose between "top_bot" (top-botton)
and "abs" (absolute)}

\item{percent_gen_select}{Percentage of genes to be selected}
}
\value{
num_gen_select Number of genes to be selected according to the percent_gen_select value
}
\description{
Checking the arguments introduces in the gene selection process.
}
\examples{
\donttest{check_gene_selection(num_genes, gen_select_type, percent_gen_select)}
}
