% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_vectors}
\alias{check_vectors}
\title{check_vectors}
\usage{
check_vectors(ncol_full_data, survival_time, survival_event, case_tag)
}
\arguments{
\item{ncol_full_data}{Number of columns (patients) of the full_data}

\item{survival_time}{Time between disease diagnosis and death (if not dead until the end of follow-up).}

\item{survival_event}{\code{logical}. Whether the patient has died or not.}

\item{case_tag}{The tag of the healthy patient (healthy or not).}
}
\value{
control_tag Return the tag of the healthy patient
}
\description{
Checking the \code{survival_time}, \code{survival_event} and \code{case_tag} introduces in the \code{GSSTDA} object.
}
\examples{
\dontrun{control_tag <- check_vectors(ncol_full_data, survival_time, survival_event, case_tag)}
}
