% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{compute_node_adjacency}
\alias{compute_node_adjacency}
\title{Computes the adjacency matrix.}
\usage{
compute_node_adjacency(nodes_list)
}
\arguments{
\item{nodes_list}{Output of the \code{levels_to_nodes} function. List of
character vectors. Each of the vectors contains the names of the
individuals at each node.}
}
\value{
It returns a matrix of magnitude n nodes x n nodes that stores a
1 if there are shared samples in two given nodes and a 0 otherwise.
}
\description{
It computes the adjacency matrix between nodes. Two nodes
are considered connected if they share at least one individual.
}
\examples{
\dontrun{
compute_node_adjacency(nodes_list)}
}
