% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Denoising.R
\name{denoise_rectangular_matrix}
\alias{denoise_rectangular_matrix}
\title{Rectangular matrix denoiser}
\usage{
denoise_rectangular_matrix(input_mat)
}
\arguments{
\item{input_mat}{A noisy matrix to denoise.}
}
\value{
Returns a same dimension denoised version of the matrix.
}
\description{
Takes a rectangular matrix composed by the addition of a signal and a gaussian noise matrix and returns a matrix of the same dimension that is denoised through a singular value truncation proccess.
}
\examples{
denoise_rectangular_matrix(matrix(c(1,2,3,4,5,2,3,1,2,3),ncol = 2))
}
