% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGSA.R
\name{fun_to_int}
\alias{fun_to_int}
\title{Marcenko-Pastur distribution to integrate.}
\usage{
fun_to_int(t, bet)
}
\arguments{
\item{t}{Parameter t}

\item{bet}{Beta value. Aspect ratio of the input matrix.  \eqn{\frac{m}{n}}, were m is the number  of rows of the input matrix and n the number of columns.}
}
\value{
returns the function value for a specific t and a particular aspect ration m/n.
}
\description{
This function is an auxiliary function includes  the marcenco-pastur function that was to be integrated to find the upper integration bound that produces and area of 1/2.
}
\examples{
fun_to_int(1,0.3)
}
