% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_selection.R
\name{gene_selection_surv}
\alias{gene_selection_surv}
\title{gene_selection_surv}
\usage{
gene_selection_surv(
  disease_component,
  p_Data,
  status_Col_Name,
  status_Value,
  cox_all,
  n_top,
  type_sel = c("Top_Bot", "Abs")
)
}
\arguments{
\item{disease_component}{A matrix with the disease component data of the entire dataset obtained by \code{generate_disease_component} function.}

\item{p_Data}{Data.frame with the phenotype data.}

\item{status_Col_Name}{Column for sample filtering.}

\item{status_Value}{Value for the sample filtering co-variate.}

\item{cox_all}{Output from the \code{cox_all_genes} function.}

\item{n_top}{Number of markers presenting the maximum values for the product.}

\item{type_sel}{Select top from bottom and top or from absolute value.}
}
\value{
vector de los genes seleccionados
}
\description{
Selects markers for mapper based on the product of standard deviation plus one times the Z score obtained through fitting a cox proportional hazard model to the expression of each gene.
}
\examples{
\dontrun{
gene_selection_surv(disease_component, p_Data, status_Col_Name, status_Value, cox_all, n_top)
}
}
