% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGSA.R
\name{generate_disease_component}
\alias{generate_disease_component}
\title{Generate disease component matrix.}
\usage{
generate_disease_component(full_data, normal_space)
}
\arguments{
\item{full_data}{Matrix containing the full dataset.}

\item{normal_space}{Denoised flattened healthy tissue data.}
}
\value{
A matrix containing the disease component of the complete dataset.
}
\description{
This functions produces a disease component matrix from the complete dataset and the denoised flattened normal data.
}
\examples{
full_data <- matrix(stats::rnorm(120),ncol=20)
normal_tissue <- full_data[,11:20]
normal_tissue_f <- flatten_normal_tiss(normal_tissue)
normal_tissue_f_d <- denoise_rectangular_matrix(normal_tissue_f)
disease_component <- generate_disease_component(full_data,normal_tissue_f_d)
}
