% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEO_Eset_Norm_frma_Collapsed}
\alias{GEO_Eset_Norm_frma_Collapsed}
\title{GEO_Eset_Norm_frma_Collapsed}
\format{
Gene expression matrix.
\describe{
\item{GEO_Eset_Norm_frma_Collapsed}{Gene expression of 104 breast cancer samples and 17 healthy tissue samples}
...
}
}
\usage{
GEO_Eset_Norm_frma_Collapsed
}
\description{
A matrix containing gene expression data of 104 breast cancer samples and 17 healthy breast tissues.
}
\keyword{datasets}
