% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapper.R
\name{get_information_from_results}
\alias{get_information_from_results}
\title{Get information from results}
\usage{
get_information_from_results(mapper_object)
}
\arguments{
\item{mapper_object}{Result object from a \code{mapper} function.}
}
\value{
A mapper_information object (a list) with informative parameters
of the mapper result, in this order: the number of nodes, the average node
size, the standard deviation of the node size, the number of connections
between nodes, the proportion of connections to all possible connections
and the number of ramifications.
}
\description{
This functions retrieves additional information about the
mapper results.
}
\examples{
\dontrun{
get_information_from_results(mapper_object)
}
}
