% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{get_intervals_One_D}
\alias{get_intervals_One_D}
\title{Extract intervals from filter function output values.}
\usage{
get_intervals_One_D(filter_values, num_intervals, percent_overlap)
}
\arguments{
\item{filter_values}{Vector obtained after applying the filtering function
to the input matrix, i.e, a vector with the filtering function
values for each included sample.}

\item{num_intervals}{Number of intervals to divide the filtering function
values in.}

\item{percent_overlap}{Percentage of overlap between intervals.}
}
\value{
Returns a list with the set of intervals for the filtering function
values.
}
\description{
It calculates the intervals of the given output values of a filter function
according to the given number and percentage of overlap.
}
\examples{
\dontrun{
get_intervals_One_D(filter_values,num_intervals,percent_overlap)}
}
