% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGSA.R
\name{get_lambda}
\alias{get_lambda}
\title{Computes lambda}
\usage{
get_lambda(bet)
}
\arguments{
\item{bet}{Beta value. Aspect ratio of the input matrix.  \deqn{\frac{m}{n}},
were m is the number of rows of the input matrix and n the number of columns.}
}
\value{
Numeric. Lambda value.
}
\description{
Computes the value of lambda as defined in: "The Optimal
Hard Threshold for Singular Values is \eqn{\sqrt(4/ 3)}".
}
\examples{
get_lambda(0.3)
}
