% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_selection.R
\name{get_survival_related_genes}
\alias{get_survival_related_genes}
\title{Gene selector based on association to survival.}
\usage{
get_survival_related_genes(cox_all, quantile_threshold = c(0.05, 0.95))
}
\arguments{
\item{cox_all}{A matrix output from the \code{cox_all_genes} function}

\item{quantile_threshold}{A two element vector indicating the bottom and top quantile for gene selection.}
}
\value{
Returns a list of genes that present z-values above or below the selected quantile thresholds.
}
\description{
Gene selector based on association to survival.
}
\examples{
\dontrun{
get_survival_related_genes(cox_all)
}
}
