% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSSTDA.R
\name{GSSTDA}
\alias{GSSTDA}
\title{GSSTDA}
\usage{
GSSTDA(GSSTDA_obj)
}
\arguments{
\item{GSSTDA_obj}{G-SS-TDA object return by \code{GSSTDA_obj} function.}
}
\value{
A \code{GSSTDA} output.
}
\description{
Gene Structure Survival using Topological Data Analysis
}
\examples{
\dontrun{
GSSTDA_obj <- GSSTDA_obj(full_data, num_intervals = 4,
                     percent_overlap = 0.5, distance_type = "euclidean",
                     num_bins_when_clustering = 8,
                     clustering_type = "hierarchical",
                     linkage_type = "single")}
GSSTDA <- GSSTDA(GSSTDA_obj)
}
