% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_SVHT_coef.R
\name{incMarPas}
\alias{incMarPas}
\title{Incomplete Marcenko-Pastur Integral}
\usage{
incMarPas(x0, beta, gamma)
}
\arguments{
\item{x0}{A numeric value representing the lower limit of the integral.}

\item{beta}{A numeric value representing the aspect ratio \eqn{m/n}, where \eqn{0 < \beta \leq 1}.}

\item{gamma}{A numeric value representing an exponent parameter.}
}
\value{
A numeric value representing the incomplete Marcenko-Pastur integral.
}
\description{
Calculates the incomplete Marcenko-Pastur integral from a lower limit to the upper bound of the distribution's support.
}
