% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{levels_to_nodes}
\alias{levels_to_nodes}
\title{Extract Information about Nodes}
\usage{
levels_to_nodes(clust_all_levels_list)
}
\arguments{
\item{clust_all_levels_list}{A list with information on the levels
obtained from the \code{clust_all_levels} function.}
}
\value{
A list including the individuals content of each detected node.
List of character vectors. Each of the vectors contains the names
of the individuals at each node.
}
\description{
Extract the nodes information based on information about
clustering. The individuals who are part of each node are identified
}
\examples{
\dontrun{
levels_to_nodes(clust_all_levels_list)}
}
