% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Funct.R
\name{lp_norm_k_powers_surv}
\alias{lp_norm_k_powers_surv}
\title{SurvPAD Filtering function}
\usage{
lp_norm_k_powers_surv(exp_matrix, p, k, cox_all)
}
\arguments{
\item{exp_matrix}{Matrix including the fit residual of the dataset and the healthy state model.}

\item{p}{Integer}

\item{k}{Integer}

\item{cox_all}{A matrix with the output of the \code{cox_all_genes} function that stores the information of all cox proportional hazard model tests for each gene in the dataset.}
}
\value{
A numeric vector including the values produced by the function for each sample in the dataset.
}
\description{
A filtering function for mapper that projects $$R$^n$ into $R$ that
}
\examples{
\dontrun{
lp_norm_k_powers(disease_state,2,1, cox_data)}

}
