% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapper.R
\name{mapper}
\alias{mapper}
\title{Mapper object}
\usage{
mapper(
  full_data,
  filter_values,
  num_intervals,
  percent_overlap,
  distance_type,
  clustering_type,
  num_bins_when_clustering,
  linkage_type,
  optimal_clustering_mode = "",
  na.rm = TRUE
)
}
\arguments{
\item{full_data}{Matrix with the columns of the input matrix
corresponding to the individuals belonging to the level.}

\item{filter_values}{Vector obtained after applying the filtering function
to the input matrix, i.e, a vector with the filtering function
values for each included sample.}

\item{num_intervals}{Number of intervals used to create the first sample
partition based on filtering values.}

\item{percent_overlap}{Percentage of overlap between intervals. Expressed
as a fraction from zero to one.}

\item{distance_type}{Type of distance to be used for clustering.
Choose between correlation ("cor") and euclidean ("euclidean"). "cor"
default option.}

\item{clustering_type}{Type of clustering method. Choose between
"hierarchical" and "PAM" (“partition around medoids”) options.
"hierarchical" default option.}

\item{num_bins_when_clustering}{Number of bins to generate the
histogram employed by the standard optimal number of cluster finder
method. Parameter not necessary if the "optimal_clust_mode" option
is "silhouette" or the "clust_type" is "PAM".}

\item{linkage_type}{Linkage criteria used in hierarchical clustering.
Choose between "single" for single-linkage clustering, "complete" for
complete-linkage clustering or "average" for average linkage clustering
(or UPGMA). Only necessary for hierarchical clustering.
"single" default option.}

\item{optimal_clustering_mode}{Method for selection optimal number of
clusters. It is only necessary if the chosen type of algorithm is
hierarchical. In this case, choose between "standard" (the method used
in the original mapper article) or "silhouette". In the case of the PAM
algorithm, the method will always be "silhouette". "silhouette" default
option.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.
If \code{FALSE}, an error occurs in case of \code{NA} rows.}
}
\value{
A \code{mapper_obj} object. It contains the values of the intervals
(interval_data), the samples included in each interval (sample_in_level),
information about the cluster to which the individuals in each interval
belong (clustering_all_levels), a list including the individuals contained
in each detected node (node_samples), their size (node_sizes), the
average of the filter function values of the individuals of each node
(node_average_filt) and the adjacency matrix linking the nodes (adj_matrix).
}
\description{
TDA are persistent homology and mapper. Persistent homology
borrows ideas from abstract algebra to identify particular aspects
related to the shape of the data such as the number of connected
components and the presence of higher-dimensional holes, whereas
mapper condenses the information of high-dimensional datasets into
a combinatory graph or simplicial complex that is referred to as
the skeleton of the dataset. This implementation is the mapper of one
dimension, i.e. using only one filter function value.
}
\examples{
\dontrun{
num_rows <- 100
full_data <- data.frame( x=2*cos(1:num_rows), y=sin(1:num_rows) )
filter_values <- list(2*cos(1:num_rows))
mapper_obj <- mapper(full_data, filter_values, num_intervals = 4,
                     percent_overlap = 0.5, distance_type = "euclidean",
                     num_bins_when_clustering = 8,
                     clustering_type = "hierarchical",
                     linkage_type = "single")}
}
