% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_SVHT_coef.R
\name{MarcenkoPasturIntegral}
\alias{MarcenkoPasturIntegral}
\title{Marcenko-Pastur Integral}
\usage{
MarcenkoPasturIntegral(x, beta)
}
\arguments{
\item{x}{A numeric value representing the upper limit of the integral.}

\item{beta}{A numeric value representing the aspect ratio \eqn{m/n}, where \eqn{0 < \beta \leq 1}.}
}
\value{
A numeric value representing the integral of the Marcenko-Pastur distribution from the lower bound to \code{x}.
}
\description{
Calculates the integral of the Marcenko-Pastur distribution from the lower bound to a specified value.
}
\examples{
x <- 0.5
beta <- 0.5
MarcenkoPasturIntegral(x, beta)
}
