% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_SVHT_coef.R
\name{MedianMarcenkoPastur}
\alias{MedianMarcenkoPastur}
\title{Median of the Marcenko-Pastur Distribution}
\usage{
MedianMarcenkoPastur(beta)
}
\arguments{
\item{beta}{A numeric value representing the aspect ratio \eqn{m/n}, where \eqn{0 < \beta \leq 1}.}
}
\value{
A numeric value representing the median of the Marcenko-Pastur distribution for the specified \code{beta}.
}
\description{
Calculates the median of the Marcenko-Pastur distribution for a given aspect ratio.
}
