% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapper.R
\name{one_D_Mapper}
\alias{one_D_Mapper}
\title{one_D_Mapper}
\usage{
one_D_Mapper(mapper_object_ini)
}
\arguments{
\item{mapper_object_ini}{Mapper TDA initializated object generated
by \code{mapper} function.}
}
\value{
A \code{mapper_obj} object. It contains the values of the intervals
(interval_data), the samples included in each interval (sample_in_level),
information about the cluster to which the individuals in each interval
belong (clustering_all_levels), a list including the individuals contained
in each detected node (node_samples), their size (node_sizes), the
average of the filter function values of the individuals of each node
(node_average_filt) and the adjacency matrix linking the nodes (adj_matrix).
Finally, information is provided on the number of nodes, the average node
size, the standard deviation of the node size, the number of connections
between nodes, the proportion of connections to all possible connections
and the number of ramifications.
}
\description{
Wrapping function to carry out the complete process.
}
\examples{
\dontrun{
one_D_Mapper(mapper_object_ini)}
}
