% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_SVHT_coef.R
\name{optimal_SVHT_coef}
\alias{optimal_SVHT_coef}
\title{Optimal Singular Value Hard Threshold (SVHT) Coefficient}
\usage{
optimal_SVHT_coef(beta, sigma_known)
}
\arguments{
\item{beta}{A numeric vector representing the aspect ratio \eqn{m/n} of the matrix to be denoised, where \eqn{0 < \beta \leq 1}.}

\item{sigma_known}{A logical value indicating if the noise level is known (TRUE) or unknown (FALSE).}
}
\value{
A numeric vector containing the optimal SVHT coefficient for each aspect ratio in \code{beta}.
}
\description{
Computes the optimal SVHT coefficient for matrix denoising based on the noise level.
}
\examples{
beta <- 0.5
sigma_known <- TRUE
optimal_SVHT_coef(beta, sigma_known)
}
