% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_SVHT_coef.R
\name{optimal_SVHT_coef_gamma_unknown}
\alias{optimal_SVHT_coef_gamma_unknown}
\title{Optimal SVHT Coefficient with Unknown Noise Level}
\usage{
optimal_SVHT_coef_gamma_unknown(beta)
}
\arguments{
\item{beta}{A numeric vector representing the aspect ratio \eqn{m/n}, where \eqn{0 < \beta \leq 1}.}
}
\value{
A numeric vector of optimal SVHT coefficients corresponding to each aspect ratio in \code{beta}.
}
\description{
Computes the optimal SVHT coefficient when the noise level is unknown.
}
\examples{
beta <- 0.5
optimal_SVHT_coef_gamma_unknown(beta)
}
