% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pData_FRMA}
\alias{pData_FRMA}
\title{pData_FRMA}
\format{
Dataframe with covariate data.
\describe{
\item{pData_FRMA}{Covariate data for the 104 breast cancer samples and 17 healthy tissue samples}
...
}
}
\usage{
pData_FRMA
}
\description{
A data.frame containing covariate data for the 104 breast cancer samples and 17 healthy breast tissues.
}
\keyword{datasets}
