% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGSA.R
\name{results_DGSA}
\alias{results_DGSA}
\title{results DGSA}
\usage{
results_DGSA(matrix_disease_component, case_tag, control_tag)
}
\arguments{
\item{matrix_disease_component}{Disease component matrix that contains
the disease component of all patients. Output of the function
\code{generate_disease_component}.}

\item{case_tag}{Character vector of the same length as the number of
columns of full_data. Patients must be in the same order as in full_data.
It must be indicated for each patient whether he/she is healthy or not.
One value should be used to indicate whether the patient is healthy and
another value should be used to indicate whether the patient's sample is
tumourous. The user will then be asked which one indicates whether
the patient is healthy. Only two values are valid in the vector in total.}

\item{control_tag}{Value of the case_tag vector corresponding to
healthy patients.}
}
\description{
It calculates the 100 genes with the highest variability in the matrix
disease component between samples and use them to draw the heat map.
}
