% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{samples_in_levels}
\alias{samples_in_levels}
\title{Samples in levels}
\usage{
samples_in_levels(interval_data, filter_values)
}
\arguments{
\item{interval_data}{Filter function intervals. List with the set of
intervals for the filtering function values produced by the
\code{get_intervals_One_D} function.}

\item{filter_values}{Vector obtained after applying the filtering
function to the input matrix, i.e, a vector with the filtering function
values for each included sample.}
}
\value{
A list of character vectors with the samples included
in each of the levels (i.e. each of the intervals of the values of the
filter functions).
}
\description{
This function returns a list of vectors containing the
individuals included at each level, i.e. the vectors of individuals with
a value of the filter function within each of the intervals.
}
\examples{
\dontrun{
samples_in_levels(interval_data,filter_values)}
}
