% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival_time}
\alias{survival_time}
\title{Survival time vector
Numeric vector of length 121 containing the time in months until
the death of the patient or until the end of the follow-up in case the
patient has not died.}
\format{
Numeric vector
\describe{
\item{survival_time}{Numeric vector of length 121 containing the time
in months until the death of the patient or until the end of the follow-up
in case the patient has not died. Samples from healthy tissue contain a
value of \code{NA}.}
}
}
\usage{
survival_time
}
\description{
Survival time vector
Numeric vector of length 121 containing the time in months until
the death of the patient or until the end of the follow-up in case the
patient has not died.
}
\keyword{datasets}
