% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_tag}
\alias{case_tag}
\title{Case-control vector
Character vector of length 121 containing the group to which each
sample belongs.}
\format{
Character vector
\describe{
\item{code_tag}{Character vector of length 121 containing the group to
which each sample belongs ("NT": control, sample from healthy tissue;
"T": case, sample from neoplastic tissue).}
}
}
\source{
The data are from the study GSE42568. Information extracted from
the file GSE42568_family.soft.gz available at
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE42568}.
}
\usage{
case_tag
}
\description{
Case-control vector
Character vector of length 121 containing the group to which each
sample belongs.
}
\keyword{datasets}
