% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_filter_values}
\alias{check_filter_values}
\title{check_filter_values}
\usage{
check_filter_values(full_data, filter_values, na.rm = TRUE)
}
\arguments{
\item{full_data}{Matrix with the columns of the input matrix
corresponding to the individuals belonging to the level. This matrix could be the
genes_disease_component.}

\item{filter_values}{Vector obtained after applying the filtering function
to the input matrix, i.e, a vector with the filtering function
values for each included sample.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.
If \code{FALSE}, an error occurs in case of \code{NA} rows.}
}
\value{
\code{filter_value} and \code{full_data} without NAN's
}
\description{
Checking the filter_values introduces in the \code{mapper} object.
}
\examples{
\donttest{check_arg_mapper(full_data, filter_values, distance_type, clustering_type, linkage_type)}
}
