% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_selection.R
\name{cox_all_genes}
\alias{cox_all_genes}
\title{Survival analysis based on gene expression levels.}
\usage{
cox_all_genes(expression_vector_disease, time_vector, event_vector)
}
\arguments{
\item{expression_vector_disease}{Expression data for disease samples}

\item{time_vector}{Vector including time to relapse or time to death information}

\item{event_vector}{Numeric vector indicating if relapse or death have been produced.}
}
\value{
A matrix with the results of the application of proportional hazard models using the expression levels of eahc gene as covariate.
}
\description{
Carries out univariate cox protportional hazard models for the expression levels of each gene included in the dataset and its link with relapse-free or overall survival.
}
\examples{
\dontrun{
cox_all_genes(expression_vector_disease,time_vector,event_vector)
}
}
