% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGSA.R
\name{denoise_rectangular_matrix}
\alias{denoise_rectangular_matrix}
\title{Rectangular Matrix Denoiser.}
\usage{
denoise_rectangular_matrix(input_mat)
}
\arguments{
\item{input_mat}{A rectangular noisy matrix to denoise.}
}
\value{
A same dimension denoised version of the matrix.
}
\description{
It takes a rectangular matrix composed by the addition of
a signal matrix and a gaussian noise matrix and returns a matrix of the same
dimension that is denoised through a Singular Value Decomposition
truncation proccess. The selection of the number of singular values is
chosen following the proposal by "The optimal hard threshold
for singular values is \eqn{\sqrt(4/ 3)}". It should be used after
the function \code{flatten_normal_tiss}.
}
\examples{
denoise_rectangular_matrix(matrix(c(1,2,3,4,5,2,3,1,2,3),ncol = 2))
}
