% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSGA.R
\name{denoise_rectangular_matrix}
\alias{denoise_rectangular_matrix}
\title{Rectangular Matrix Denoiser.}
\usage{
denoise_rectangular_matrix(matrix_flatten_normal_tiss, gamma)
}
\arguments{
\item{matrix_flatten_normal_tiss}{A rectangular noisy matrix to denoise. It is return by
\code{flatten_normal_tiss} function.}

\item{gamma}{A parameter that indicates the magnitude of the noise assumed in
the flat data matrix for the generation of the Healthy State Model. If it
takes the value \code{NA} the magnitude of the noise is assumed to be unknown.
By default gamma is unknown.}
}
\value{
A the normal space which has the same dimension denoised version of the matrix
returned by \code{flatten_normal_tiss}.
}
\description{
It takes a rectangular matrix composed by the addition of
a signal matrix and a Gaussian noise matrix and returns a matrix of the same
dimension that is denoised through a Singular Value Decomposition
truncation process. The selection of the number of singular values is
chosen following the proposal by "The optimal hard threshold
for singular values is \eqn{\sqrt(4/ 3)}". It should be used after
the function \code{flatten_normal_tiss}.
}
