% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGSA.R
\name{denoise_rectangular_matrix}
\alias{denoise_rectangular_matrix}
\title{Rectangular Matrix Denoiser.}
\usage{
denoise_rectangular_matrix(matrix_flatten_normal_tiss)
}
\arguments{
\item{matrix_flatten_normal_tiss}{A rectangular noisy matrix to denoise. It is return by
\code{flatten_normal_tiss} function.}
}
\value{
A the normal space which has the same dimension denoised version of the matrix
returned by \code{flatten_normal_tiss}.
}
\description{
It takes a rectangular matrix composed by the addition of
a signal matrix and a Gaussian noise matrix and returns a matrix of the same
dimension that is denoised through a Singular Value Decomposition
truncation process. The selection of the number of singular values is
chosen following the proposal by "The optimal hard threshold
for singular values is \eqn{\sqrt(4/ 3)}". It should be used after
the function \code{flatten_normal_tiss}.
}
\examples{
\donttest{
denoise_rectangular_matrix(matrix(c(1,2,3,4,5,2,3,1,2,3),ncol = 2))
}
}
