% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_data}
\alias{full_data}
\title{full_data}
\format{
Gene expression matrix.
\describe{
\item{full_data}{Gene expression of 104 breast cancer samples and 17 healthy tissue samples}
...
}
}
\usage{
full_data
}
\description{
A matrix containing gene expression data of 104 breast cancer samples and 17 healthy breast tissues.
}
\keyword{datasets}
