% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Denoising.R
\name{get_lambda}
\alias{get_lambda}
\title{Computes lambda}
\usage{
get_lambda(bet)
}
\arguments{
\item{bet}{numeric}
}
\value{
A numeric lambda value.
}
\description{
Computes the value of lambda as defined in: The Optimal Hard Threshold for Singular Values is \eqn{\sqrt(4/ 3)}
}
\examples{
get_lambda(0.3)
}
