% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSSTDA.R
\name{GSSTDA}
\alias{GSSTDA}
\title{GSSTDA}
\usage{
GSSTDA(
  full_data,
  survival_time,
  survival_event,
  case_tag,
  num_intervals,
  percent_overlap,
  distance_type,
  clustering_type,
  num_bins_when_clustering,
  linkage_type,
  na.rm = TRUE
)
}
\arguments{
\item{full_data}{Matrix with the columns of the input matrix
corresponding to the individuals belonging to the level.}

\item{survival_time}{Time between disease diagnosis and death (if not dead until the end of follow-up).}

\item{survival_event}{\code{logical}. Whether the patient has died or not.}

\item{case_tag}{The tag of the healthy patient (healthy or not).}

\item{num_intervals}{Number of intervals used to create the first sample
partition based on filtering values.}

\item{percent_overlap}{Percentage of overlap between intervals. Expressed
as a fraction from zero to one.}

\item{distance_type}{Type of distance to be used for clustering.
Choose between correlation ("cor") and euclidean ("euclidean"). "cor"
default option.}

\item{clustering_type}{Type of clustering method. Choose between
"hierarchical" and "PAM" (“partition around medoids”) options.
"hierarchical" default option.}

\item{num_bins_when_clustering}{Number of bins to generate the
histogram employed by the standard optimal number of cluster finder
method. Parameter not necessary if the "optimal_clust_mode" option
is "silhouette" or the "clust_type" is "PAM".}

\item{linkage_type}{Linkage criteria used in hierarchical clustering.
Choose between "single" for single-linkage clustering, "complete" for
complete-linkage clustering or "average" for average linkage clustering
(or UPGMA). Only necessary for hierarchical clustering.
"single" default option.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.
If \code{FALSE}, an error occurs in case of \code{NA} rows.}
}
\value{
A \code{GSSTDA} object.
}
\description{
Gene Structure Survival using Topological Data Analysis
}
\examples{
\dontrun{
num_rows <- 100
full_data <- data.frame( x=2*cos(1:num_rows), y=sin(1:num_rows) )
filter_values <- list(2*cos(1:num_rows))
GSSTDA <- GSSTDA(full_data,  survival_time, survival_event, case_tag, num_intervals = 4,
                     percent_overlap = 0.5, distance_type = "euclidean",
                     num_bins_when_clustering = 8,
                     clustering_type = "hierarchical",
                     linkage_type = "single")}
}
