% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival_event}
\alias{survival_event}
\title{Survival event vector
Character vector of length 121 containing whether or not the patient
is deceased.}
\format{
Character vector
\describe{
\item{survival_event}{Character vector of length 121 containing whether
or not the patient is deceased ("0": no, "1": yes). Samples from healthy
tissue contain a value of \code{NA}.}
}
}
\source{
The data are from the study GSE42568. Information extracted from
the file GSE42568_family.soft.gz available at
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE42568}.
}
\usage{
survival_event
}
\description{
Survival event vector
Character vector of length 121 containing whether or not the patient
is deceased.
}
\keyword{datasets}
