% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival_event}
\alias{survival_event}
\title{survival_event}
\format{
Dataframe with covariate data.
\describe{
\item{survival_event}{Covariate data for the 104 breast cancer samples and 17 healthy tissue samples}
...
}
}
\usage{
survival_event
}
\description{
A data.frame containing covariate data for the 104 breast cancer samples and 17 healthy breast tissues.
}
\keyword{datasets}
